# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "job-bot",
    "version": "dev"
  },
  "paths": {
...
    "/api/offers/{offer_id}/status": {
      "patch": {
        "tags": [
          "applications"
        ],
        "summary": "Set Offer Status",
        "operationId": "setOfferStatus",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferStatusIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/offers/{offer_id}/expiry": {
      "patch": {
        "tags": [
          "applications"
        ],
        "summary": "Set Offer Expiry",
        "description": "« Signaler comme expirée » ou « Pas expirée » (docs/11 §1) : le choix de Kevin passe\navant la détection automatique, dans les deux sens.",
        "operationId": "setOfferExpiry",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferExpiryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferExpiryOut"
                }
              }
            }
          },
          "409": {
            "description": "Candidature déjà envoyée"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/offers/{offer_id}/address": {
      "patch": {
        "tags": [
          "applications"
        ],
        "summary": "Set Offer Address",
        "description": "Adresse corrigée ou saisie par Kevin : prioritaire sur toute source automatique.",
        "operationId": "setOfferAddress",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferAddressIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfferAddressOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/offers/{offer_id}/application-prefill": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Get Application Prefill",
        "description": "Valeurs proposées pour « Marquer comme envoyée ».",
        "operationId": "getApplicationPrefill",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationPrefill"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "List Applications",
        "operationId": "listApplications",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApplicationOut"
                  },
                  "title": "Response Listapplications"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "applications"
        ],
        "summary": "Create Application",
        "operationId": "createApplication",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "409": {
            "description": "Candidature déjà enregistrée pour cette offre"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/summary": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Get Applications Summary",
        "operationId": "getApplicationsSummary",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^\\d{4}-\\d{2}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{application_id}": {
      "put": {
        "tags": [
          "applications"
        ],
        "summary": "Update Application",
        "operationId": "updateApplication",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplicationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "applications"
        ],
        "summary": "Delete Application",
        "description": "Annule une candidature enregistrée par erreur ; l'offre revient « en préparation ».",
        "operationId": "deleteApplication",
        "parameters": [
          {
            "name": "application_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Application Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/identity": {
      "get": {
        "tags": [
          "applications"
        ],
        "summary": "Get Identity",
        "operationId": "getIdentity",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "applications"
        ],
        "summary": "Put Identity",
        "operationId": "saveIdentity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Identity"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Identity"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApplicationMethod": {
        "type": "string",
        "enum": [
          "electronique",
          "ecrit",
          "telephone",
          "personnel"
        ],
        "title": "ApplicationMethod",
        "description": "Mode de candidature, au sens du formulaire ORP."
      },
...
      "ApplicationPrefill": {
        "properties": {
          "sent_at": {
            "type": "string",
            "format": "date",
            "title": "Sent At"
          },
          "method": {
            "$ref": "#/components/schemas/ApplicationMethod",
            "default": "electronique"
          },
          "assigned_by_orp": {
            "type": "boolean",
            "title": "Assigned By Orp",
            "default": false
          },
          "company": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Company"
          },
          "company_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Address"
          },
          "contact_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Name"
          },
          "contact_phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "job_title": {
            "type": "string",
            "maxLength": 300,
            "minLength": 1,
            "title": "Job Title"
          },
          "location": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "rate_text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Rate Text"
          },
          "offer_id": {
            "type": "integer",
            "title": "Offer Id"
          }
        },
        "type": "object",
        "required": [
          "sent_at",
          "company",
          "job_title",
          "offer_id"
        ],
        "title": "ApplicationPrefill"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MonthSummary": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "target": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          }
        },
        "type": "object",
        "required": [
          "month",
          "count",
          "target"
        ],
        "title": "MonthSummary"
      },
...
      "OfferAddressIn": {
        "properties": {
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 300
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          }
        },
        "type": "object",
        "title": "OfferAddressIn"
      },
      "OfferAddressOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "company_address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Address"
          },
          "company_address_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company Address Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_address",
          "company_address_source"
        ],
        "title": "OfferAddressOut"
      },
...
      "OfferExpiryIn": {
        "properties": {
          "expired": {
            "type": "boolean",
            "title": "Expired"
          }
        },
        "type": "object",
        "required": [
          "expired"
        ],
        "title": "OfferExpiryIn"
      },
      "OfferExpiryOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "expired_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expired At"
          },
          "expiry_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Source"
          },
          "expiry_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Override"
          }
        },
        "type": "object",
        "required": [
          "id",
          "expired_at",
          "expiry_source",
          "expiry_override"
        ],
        "title": "OfferExpiryOut"
      },
...
      "OfferStatus": {
        "type": "string",
        "enum": [
          "new",
          "filtered_out",
          "to_review",
          "later",
          "ignored",
          "preparing",
          "applied"
        ],
        "title": "OfferStatus"
      },
      "OfferStatusIn": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "to_review",
              "later",
              "ignored",
              "preparing"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OfferStatusIn"
      },
      "OfferStatusOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/OfferStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "title": "OfferStatusOut"
      },
...
    }
  }
}